MediaAutomation Server/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
python-jose==3.3.0
passlib[bcrypt]==1.7.4
obs-websocket-py==0.6.3
httpx==0.24.1
APScheduler==3.10.1
python-multipart==0.0.6
